Make ng::side_engine::get_color()'s argument a signed integer
Fixes a signed vs. unsigned type comparison warning.

+++ src/game_initialization/connect_engine.cpp
@@ -1344,7 +1344,7 @@ std::vector<std::string> side_engine::get_colors() const
 	return res;
 }
 
-std::string side_engine::get_color(size_t index) const
+std::string side_engine::get_color(int index) const
 {
 	if(index == -1) {
 		index = color();
